+++ src/components/files/ViewFolder.js
@@ -1,7 +1,7 @@
-import React, { useEffect, Fragment } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { Link, Redirect, useHistory } from "react-router-dom";
-import moment from "moment";
+import React, { useEffect, Fragment } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { Link, Redirect, useHistory } from 'react-router-dom';
+import moment from 'moment';
 import {
   Table,
   TableBody,
@@ -13,14 +13,15 @@ import {
   Breadcrumbs,
   Typography,
   Chip,
-} from "@material-ui/core/";
-import { CloudDownload, CloudUpload } from "@material-ui/icons";
-import NavigateNextIcon from "@material-ui/icons/NavigateNext";
-import FolderIcon from "@material-ui/icons/Folder";
-import Circular from "../layout/Circular";
-import { getFolders } from "../../actions/folderActions";
-import useStyles from "./StyleFiles";
-import Hidden from "@material-ui/core/Hidden";
+} from '@material-ui/core/';
+import { CloudDownload, CloudUpload } from '@material-ui/icons';
+import NavigateNextIcon from '@material-ui/icons/NavigateNext';
+import FolderIcon from '@material-ui/icons/Folder';
+import Circular from '../layout/Circular';
+import { getFolders } from '../../actions/folderActions';
+import useStyles from '../../style/StyleFiles';
+import Hidden from '@material-ui/core/Hidden';
+import SignIn from "../authen/SignIn";
 const ViewFolder = () => {
   const classes = useStyles();
   const history = useHistory();
@@ -34,72 +35,77 @@ const ViewFolder = () => {
       : dispatch(getFolders(authdata.user_id));
   }, []);
   if (loading) {
-    console.log("loading >>> " + loading);
+    console.log('loading >>> ' + loading);
   }
 
   return (
     <div>
       {authenticated && (
         <div>
-          {authdata.authorized_id == 1 ? (
-            <Fragment>
-              <Grid container container className={classes.gridContainer}>
-                <Paper className={classes.paper}>
-                  <Grid container>
-                    <Breadcrumbs
-                      className={classes.breadcrumbs}
-                      separator={<NavigateNextIcon />}
-                      aria-label="breadcrumb"
-                    >
-                      <Typography className={classes.text} color="textPrimary">
-                        โฟลเดอร์ทั้งหมด
-                      </Typography>
-                    </Breadcrumbs>
-                  </Grid>
-                </Paper>
-                <Paper className={classes.paper}>
-                  <Hidden smDown>
-                    <Table className={classes.table}>
-                      <TableHead>
-                        <TableCell className={classes.tableCellName}>
-                          <Typography
-                            color="textPrimary"
-                            className={classes.text}
-                          >
-                            ชื่อโฟลเดอร์
-                          </Typography>
-                        </TableCell>
-                        <TableCell align="center">
-                          <Typography
-                            color="textPrimary"
-                            className={classes.text}
-                          >
-                            วันที่แก้ไข
-                          </Typography>
-                        </TableCell>
-                        <TableCell align="center" colSpan={2}>
-                          <Typography
-                            color="textPrimary"
-                            className={classes.text}
-                          >
-                            สิทธิ์การใช้งานโฟลเดอร์
-                          </Typography>
-                        </TableCell>
-                      </TableHead>
-                      <TableBody>
-                        {!loading && folders !== null
-                          ? folders.map((folder) => (
+          {
+            authdata.authorized_id == 1 ? (
+              <Fragment>
+                <Grid container container className={classes.gridContainer}>
+                  <Paper className={classes.paper}>
+                    <Grid container>
+                      <Breadcrumbs
+                        className={classes.breadcrumbs}
+                        separator={<NavigateNextIcon />}
+                        aria-label='breadcrumb'
+                      >
+                        <Typography
+                          className={classes.text}
+                          color='textPrimary'
+                        >
+                          โฟลเดอร์ทั้งหมด
+                        </Typography>
+                      </Breadcrumbs>
+                    </Grid>
+                  </Paper>
+                  <Paper className={classes.paper}>
+                    <Hidden smDown>
+                      <Table className={classes.table}>
+                        <TableHead>
+                          <TableCell className={classes.tableCellName}>
+                            <Typography
+                              color='textPrimary'
+                              className={classes.text}
+                            >
+                              ชื่อโฟลเดอร์
+                            </Typography>
+                          </TableCell>
+                          <TableCell align='center'>
+                            <Typography
+                              color='textPrimary'
+                              className={classes.text}
+                            >
+                              วันที่แก้ไข
+                            </Typography>
+                          </TableCell>
+                          <TableCell align='center' colSpan={2}>
+                            <Typography
+                              color='textPrimary'
+                              className={classes.text}
+                            >
+                              สิทธิ์การใช้งานโฟลเดอร์
+                            </Typography>
+                          </TableCell>
+                        </TableHead>
+                        <TableBody>
+                          {!loading && folders !== null
+                            ? folders.map((folder) => (
                               <TableRow key={folder.folder_id}>
                                 <TableCell>
                                   <Link
                                     to={{
                                       pathname:
-                                        "/ViewFiles/" + folder.folder_id,
+                                        '/ViewFiles/' + folder.folder_id,
                                       state: {
                                         folder_id: folder.folder_id,
                                         folder_name: folder.folder_name,
                                         access_upload: folder.access_upload,
-                                        access_download: folder.access_download,
+                                        access_download:
+                                          folder.access_download,
                                       },
                                     }}
                                   >
@@ -116,83 +122,84 @@ const ViewFolder = () => {
                                     </Grid>
                                   </Link>
                                 </TableCell>
-                                <TableCell align="center">
+                                <TableCell align='center '>
                                   <Typography
-                                    color="textPrimary"
+                                    color='textPrimary'
                                     className={classes.text}
                                   >
                                     {moment
                                       .utc(folder.folder_updated)
-                                      .add(3, "minutes")
-                                      .format("DD-MM-YYYY HH:mm")}
+                                      .add(3, 'minutes')
+                                      .format('DD-MM-YYYY HH:mm')}
                                   </Typography>
                                 </TableCell>
-                                <TableCell align="right">
+                                <TableCell align='right'>
                                   {folder.access_download ? (
                                     <Chip
                                       icon={<CloudDownload />}
-                                      label="Download"
-                                      color="secondary"
+                                      label='Download'
+                                      color='secondary'
                                       style={{
-                                        backgroundColor: "#4CAF50",
-                                        width: "110px",
+                                        backgroundColor: '#4CAF50',
+                                        width: '110px',
                                       }}
                                     />
                                   ) : (
-                                    <Chip
-                                      disabled={true}
-                                      icon={<CloudDownload />}
-                                      label="Download"
-                                      style={{
-                                        width: "110px",
-                                      }}
-                                    />
-                                  )}
+                                      <Chip
+                                        disabled={true}
+                                        icon={<CloudDownload />}
+                                        label='Download'
+                                        style={{
+                                          width: '110px',
+                                        }}
+                                      />
+                                    )}
                                 </TableCell>
-                                <TableCell align="left">
+                                <TableCell align='left'>
                                   {folder.access_upload ? (
                                     <Chip
                                       icon={<CloudUpload />}
-                                      label="Upload"
-                                      color="secondary"
+                                      label='Upload'
+                                      color='secondary'
                                       style={{
-                                        backgroundColor: "#2196F3",
-                                        width: "110px",
+                                        backgroundColor: '#2196F3',
+                                        width: '110px',
                                       }}
                                     />
                                   ) : (
-                                    <Chip
-                                      disabled={true}
-                                      icon={<CloudUpload />}
-                                      label="Upload"
-                                      style={{
-                                        width: "110px",
-                                      }}
-                                    />
-                                  )}
+                                      <Chip
+                                        disabled={true}
+                                        icon={<CloudUpload />}
+                                        label='Upload'
+                                        style={{
+                                          width: '110px',
+                                        }}
+                                      />
+                                    )}
                                 </TableCell>
                               </TableRow>
                             ))
-                          : console.log("Nodata")}
-                      </TableBody>
-                    </Table>
-                  </Hidden>
-                  <Hidden mdUp>
-                    <Table>
-                      <TableBody>
-                        {!loading && folders !== null
-                          ? folders.map((folder) => (
+                            : console.log('Nodata')}
+                        </TableBody>
+                      </Table>
+                    </Hidden>
+                    <Hidden mdUp>
+                      <Table>
+                        <TableBody>
+                          {!loading && folders !== null
+                            ? folders.map((folder) => (
                               <TableRow key={folder.folder_id}>
                                 <TableCell>
                                   <Link
                                     to={{
                                       pathname:
-                                        "/ViewFiles/" + folder.folder_id,
+                                        '/ViewFiles/' + folder.folder_id,
                                       state: {
                                         folder_id: folder.folder_id,
                                         folder_name: folder.folder_name,
                                         access_upload: folder.access_upload,
-                                        access_download: folder.access_download,
+                                        access_download:
+                                          folder.access_download,
                                       },
                                     }}
                                   >
@@ -216,71 +223,73 @@ const ViewFolder = () => {
                                         >
                                           {moment
                                             .utc(folder.folder_updated)
-                                            .add(3, "minutes")
-                                            .format("DD-MM-YYYY HH:mm")}
+                                            .add(3, 'minutes')
+                                            .format('DD-MM-YYYY HH:mm')}
                                         </Typography>
                                       </Grid>
                                     </Grid>
                                   </Link>
                                 </TableCell>
                                 <TableCell>
+
                                   {folder.access_download ? (
-                                    <CloudDownload
-                                      color="secondary"
+                                    <CloudDownload color='secondary'
                                       style={{
-                                        color: "#4CAF50",
-                                        width: "30px",
-                                      }}
-                                    />
+                                        color: '#4CAF50',
+                                        width: '30px',
+                                      }} />
+
                                   ) : (
-                                    <CloudDownload
-                                      color="disabled"
-                                      style={{
-                                        width: "30px",
-                                      }}
-                                    />
-                                  )}
+                                      <CloudDownload
+                                        color='disabled'
+                                        style={{
+                                          width: '30px',
+                                        }} />
+                                    )}
 
                                   {folder.access_upload ? (
                                     <CloudUpload
                                       style={{
-                                        color: "#2196F3",
-                                        width: "30px",
-                                      }}
-                                    />
+                                        color: '#2196F3',
+                                        width: '30px',
+                                      }} />
+
                                   ) : (
-                                    <CloudUpload
-                                      color="disabled"
-                                      style={{
-                                        width: "30px",
-                                      }}
-                                    />
-                                  )}
+                                      <CloudUpload
+                                        color="disabled"
+                                        style={{
+                                          width: '30px',
+                                        }} />
+
+                                    )}
+
                                 </TableCell>
                               </TableRow>
                             ))
-                          : console.log("Nodata")}
-                      </TableBody>
-                    </Table>
-                  </Hidden>
-                  {folders === null && (
-                    <Table>
-                      <TableRow>
-                        <TableCell className={classes.emptyTable}>
-                          <Typography>
-                            {" ไม่พบโฟลเดอร์ที่มีสิทธิ์เข้าถึง "}
-                          </Typography>
-                        </TableCell>
-                      </TableRow>
-                    </Table>
-                  )}
-                  {loading && <Circular />}
-                </Paper>
-              </Grid>
-            </Fragment>
-          ) : (
-            <Redirect to={"/viewfolderadmin"} />
-          )}
+                            : console.log('Nodata')}
+                        </TableBody>
+                      </Table>
+                    </Hidden>
+                    {folders === null && (
+                      <Table>
+                        <TableRow>
+                          <TableCell className={classes.emptyTable}>
+                            <Typography>
+                              {' ไม่พบโฟลเดอร์ที่มีสิทธิ์เข้าถึง '}
+                            </Typography>
+                          </TableCell>
+                        </TableRow>
+                      </Table>
+                    )}
+                    {loading && <Circular />}
+                  </Paper>
+                </Grid>
+              </Fragment>
+            ) : (
+                <Redirect to={'/viewfolderadmin'} />
+              )
+            //history.push('/viewfolderamin')
+          }
         </div>
       )}
     </div>
