+++ src/components/files/ViewFolderAdmin.js
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from "react-redux";
 import { Link, Redirect } from "react-router-dom";
 import NavigateNextIcon from "@material-ui/icons/NavigateNext";
 import FolderIcon from "@material-ui/icons/Folder";
-import AddFolder from "./AddFolder";
-import useStyles from "./StyleFiles";
-import MenuFolder from "./MenuFolder";
+import AddFolder from "../modal/AddFolder";
+import useStyles from "../../style/StyleFiles";
+import MenuFolder from "../menu/MenuFolder";
 import Circular from "../layout/Circular";
 import { useSnackbar } from "notistack";
 import moment from "moment";
@@ -122,56 +122,56 @@ const ViewFolderAdmin = (props) => {
                       <TableBody>
                         {!loading && foldersadmin != null
                           ? foldersadmin.map((folder) => (
-                              <TableRow key={folder.folder_id}>
-                                <TableCell>
-                                  <Link
-                                    to={{
-                                      pathname:
-                                        "/viewfilesadmin/" + folder.folder_id,
-                                      state: {
-                                        folder_id: folder.folder_id,
-                                        folder_name: folder.folder_name,
-                                      },
-                                    }}
+                            <TableRow key={folder.folder_id}>
+                              <TableCell>
+                                <Link
+                                  to={{
+                                    pathname:
+                                      "/viewfilesadmin/" + folder.folder_id,
+                                    state: {
+                                      folder_id: folder.folder_id,
+                                      folder_name: folder.folder_name,
+                                    },
+                                  }}
+                                >
+                                  <Grid
+                                    container
+                                    className={classes.iconAlign}
                                   >
-                                    <Grid
-                                      container
-                                      className={classes.iconAlign}
-                                    >
-                                      <Grid>
-                                        <FolderIcon
-                                          className={classes.iconFolderTable}
-                                        />
-                                      </Grid>
-                                      <Grid>
-                                        <Typography className={classes.text}>
-                                          {folder.folder_name}
-                                        </Typography>
-                                      </Grid>
+                                    <Grid>
+                                      <FolderIcon
+                                        className={classes.iconFolderTable}
+                                      />
                                     </Grid>
-                                  </Link>
-                                </TableCell>
-                                <TableCell align="center">
-                                  <Typography className={classes.text}>
-                                    {moment
-                                      .utc(folder.folder_updated)
-                                      .add(3, "minutes")
-                                      .format("DD-MM-YYYY HH:mm")}
-                                  </Typography>
-                                </TableCell>
-                                <TableCell align="center">
-                                  <MenuFolder
-                                    delete={handleDeleteFolder}
-                                    edit={handleChangeFolderName}
-                                    snackAlert={snackAlert}
-                                    refresh={refresh}
-                                    folder_name_old={folder.folder_name}
-                                    folder_name={folder.folder_name}
-                                    folder_id={folder.folder_id}
-                                  />
-                                </TableCell>
-                              </TableRow>
-                            ))
+                                    <Grid>
+                                      <Typography className={classes.text}>
+                                        {folder.folder_name}
+                                      </Typography>
+                                    </Grid>
+                                  </Grid>
+                                </Link>
+                              </TableCell>
+                              <TableCell align="center">
+                                <Typography className={classes.text}>
+                                  {moment
+                                    .utc(folder.folder_updated)
+                                    .add(3, "minutes")
+                                    .format("DD-MM-YYYY HH:mm")}
+                                </Typography>
+                              </TableCell>
+                              <TableCell align="center">
+                                <MenuFolder
+                                  delete={handleDeleteFolder}
+                                  edit={handleChangeFolderName}
+                                  snackAlert={snackAlert}
+                                  refresh={refresh}
+                                  folder_name_old={folder.folder_name}
+                                  folder_name={folder.folder_name}
+                                  folder_id={folder.folder_id}
+                                />
+                              </TableCell>
+                            </TableRow>
+                          ))
                           : console.log("Nodata")}
                       </TableBody>
                     </Table>
@@ -181,58 +181,58 @@ const ViewFolderAdmin = (props) => {
                       <TableBody>
                         {!loading && foldersadmin != null
                           ? foldersadmin.map((folder) => (
-                              <TableRow key={folder.folder_id}>
-                                <TableCell>
-                                  <Link
-                                    to={{
-                                      pathname:
-                                        "/viewfilesadmin/" + folder.folder_id,
-                                      state: {
-                                        folder_id: folder.folder_id,
-                                        folder_name: folder.folder_name,
-                                      },
-                                    }}
+                            <TableRow key={folder.folder_id}>
+                              <TableCell>
+                                <Link
+                                  to={{
+                                    pathname:
+                                      "/viewfilesadmin/" + folder.folder_id,
+                                    state: {
+                                      folder_id: folder.folder_id,
+                                      folder_name: folder.folder_name,
+                                    },
+                                  }}
+                                >
+                                  <Grid
+                                    container
+                                    className={classes.iconAlign}
                                   >
-                                    <Grid
-                                      container
-                                      className={classes.iconAlign}
-                                    >
-                                      <Grid>
-                                        <FolderIcon
-                                          className={classes.iconFolderTable}
-                                        />
-                                      </Grid>
-                                      <Grid>
-                                        <Typography
-                                          className={classes.nowrapText}
-                                        >
-                                          {folder.folder_name}
-                                        </Typography>
-                                        <Typography
-                                          className={classes.textDate}
-                                        >
-                                          {moment
-                                            .utc(folder.folder_updated)
-                                            .add(3, "minutes")
-                                            .format("DD-MM-YYYY HH:mm")}
-                                        </Typography>
-                                      </Grid>
+                                    <Grid>
+                                      <FolderIcon
+                                        className={classes.iconFolderTable}
+                                      />
                                     </Grid>
-                                  </Link>
-                                </TableCell>
-                                <TableCell align="center">
-                                  <MenuFolder
-                                    delete={handleDeleteFolder}
-                                    edit={handleChangeFolderName}
-                                    snackAlert={snackAlert}
-                                    refresh={refresh}
-                                    folder_name_old={folder.folder_name}
-                                    folder_name={folder.folder_name}
-                                    folder_id={folder.folder_id}
-                                  />
-                                </TableCell>
-                              </TableRow>
-                            ))
+                                    <Grid>
+                                      <Typography
+                                        className={classes.nowrapText}
+                                      >
+                                        {folder.folder_name}
+                                      </Typography>
+                                      <Typography
+                                        className={classes.textDate}
+                                      >
+                                        {moment
+                                          .utc(folder.folder_updated)
+                                          .add(3, "minutes")
+                                          .format("DD-MM-YYYY HH:mm")}
+                                      </Typography>
+                                    </Grid>
+                                  </Grid>
+                                </Link>
+                              </TableCell>
+                              <TableCell align="center">
+                                <MenuFolder
+                                  delete={handleDeleteFolder}
+                                  edit={handleChangeFolderName}
+                                  snackAlert={snackAlert}
+                                  refresh={refresh}
+                                  folder_name_old={folder.folder_name}
+                                  folder_name={folder.folder_name}
+                                  folder_id={folder.folder_id}
+                                />
+                              </TableCell>
+                            </TableRow>
+                          ))
                           : console.log("Nodata")}
                       </TableBody>
                     </Table>
@@ -248,8 +248,8 @@ const ViewFolderAdmin = (props) => {
                       </TableRow>
                     </Table>
                   ) : (
-                    console.log("folder empty")
-                  )}
+                      console.log("folder empty")
+                    )}
                   {loading && (
                     <div className={classes.loading}>
                       <Circular />
@@ -260,8 +260,8 @@ const ViewFolderAdmin = (props) => {
               </Grid>
             </Fragment>
           ) : (
-            <Redirect to="/" />
-          )}
+              <Redirect to="/" />
+            )}
         </div>
       )}
     </div>
