+++ src/components/modal/ConfirmDeleteFolder.js
@@ -1,11 +1,9 @@
 import React from "react";
-import "./upload_style.css";
-import "./style.css";
 import Modal from "@material-ui/core/Modal";
 import Backdrop from "@material-ui/core/Backdrop";
 import Fade from "@material-ui/core/Fade";
 import { Button, Typography } from "@material-ui/core";
-import useStyles from "./StyleFiles";
+import useStyles from "../../style/StyleFiles";
 import RemoveIcon from "@material-ui/icons/Remove";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import { MenuItem } from "@material-ui/core/";
