+++ src/components/modal/ConfirmDowloadAdmin.js
@@ -1,33 +1,31 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import "./upload_style.css";
-import "./style.css";
 import GetAppIcon from "@material-ui/icons/GetApp";
 import {
   Button,
   Typography,
-  IconButton,
+  MenuItem,
+  ListItemIcon,
   Fade,
   Backdrop,
   Modal,
 } from "@material-ui/core";
-import useStyles from "./StyleFiles";
+import useStyles from "../../style/StyleFiles";
 import { download } from "../../actions/fileActions";
-import Hidden from "@material-ui/core/Hidden";
 
-export default function ConfirmDownload(props) {
+export default function ConfirmDownloadAdmin(props) {
   const dispatch = useDispatch();
   const classes = useStyles();
   const [open, setOpen] = React.useState(false);
-  const filename = props.filename;
-  const fileId = props.fileid;
+  const file = props.file;
 
   const handleDownload = () => {
-    dispatch(download(fileId, filename));
+    dispatch(download(file.file_id, file.file_name));
     handleClose();
   };
 
   const handleOpen = () => {
+    props.closeMenu();
     setOpen(true);
   };
   const handleClose = () => {
@@ -35,17 +33,14 @@ export default function ConfirmDownload(props) {
   };
   return (
     <div>
-      <Hidden xsDown>
-        <IconButton className={classes.tableMargin} onClick={handleOpen}>
+      <MenuItem onClick={handleOpen}>
+        <ListItemIcon>
           <GetAppIcon />
-        </IconButton>
-      </Hidden>
-      <Hidden smUp>
-        <IconButton className={classes.iconButton} onClick={handleOpen}>
-          <GetAppIcon />
-        </IconButton>
-      </Hidden>
-
+        </ListItemIcon>
+        <Typography variant="inherit" className={classes.menuItem}>
+          ดาวน์โหลด
+        </Typography>
+      </MenuItem>
       <Modal
         className={classes.modal}
         open={open}
@@ -59,9 +54,8 @@ export default function ConfirmDownload(props) {
         <Fade in={open}>
           <div className={classes.modalPaper}>
             <div className={classes.root}>
-              <h1>{props.count}</h1>
               <Typography className={classes.text}>
-                คุณต้องการดาวน์โหลด {props.filename} ?
+                คุณต้องการดาวน์โหลด {file.file_name} ?
               </Typography>
               <div className={classes.modalBtn}>
                 <Button
