--- src/components/styles/table.js
@@ -1,12 +1,0 @@
-import { createMuiTheme } from '@material-ui/core/styles';
-
-export const theme = createMuiTheme({
-    overrides: {
-        MuiTableCell: {
-            root: {  //This can be referred from Material UI API documentation. 
-                padding: '100px 100px',
-                backgroundColor: "#eaeaea",
-            },
-        },
-    },
-});
